requirement.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
django-filter==24.3
django-multiselectfield==0.1.13
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.8.2
